requirements.txt
PyQt5>=5.15.4
PyQt-Fluent-Widgets>=1.8.6
opencv-python
numpy
pillow
Nuitka